+++ sgrun/src/sgrun/sg_scheduler.h
@@ -4,6 +4,7 @@
 #include "core/vec.h"
 #include <stdbool.h>
 #include <stddef.h>
+#include <stdint.h>
 
 typedef bool(*PollFutFn)(void*);
 typedef bool(*StartFutFn)(void*);
@@ -17,13 +18,13 @@ typedef struct {
 typedef struct {
   vec(SGFuture) pending;
   vec(SGFuture) executing;
-  size_t max_jobs;
+  uint32_t max_jobs;
   /// id of current job
-  size_t curr;
+  uint32_t curr;
 } SGJScheduler;
 
 
-SGJScheduler sgjs_new(size_t max_jobs);
+SGJScheduler sgjs_new(uint32_t max_jobs);
 
 /// Await all the jobs added, blocks till all jobs are completed
 void sgjs_await(SGJScheduler* sched);
